Moved New Opportunity form to be a modal on the main dashboard page.
