+++ client/src/components/NewOpp.jsx
@@ -5,7 +5,7 @@ import Cookies from "js-cookie";
 
 
 function NewOpportunityForm(props) {
-  const {user} = props;
+  const {user, afterSubmit} = props;
   const [formData, setFormData] = useState({
     opportunity_name: '',
     prospect_name: '',
@@ -36,14 +36,14 @@ function NewOpportunityForm(props) {
     })
       .then(res => {
         console.log("Opportunity created: ", res.data);
-        navigate("/dashboard")
+        if (afterSubmit) afterSubmit();
       })
       .catch(err => console.log(err));
   };  
 
   return (
     <div className="container">
-      <h1>Create a New Opportunity</h1>
+      {/* <h1>Create a New Opportunity</h1> */}
       <form onSubmit={handleSubmit}>
         <div className="mb-3">
           <label className="form-label">Opportunity Name</label>
